test(text): migrate
